add in new tests and refactor to TS
